ENH: associate sprite flippnig and animatinos
- A simple Move method can be called to move the customer towards a goal
