--- Assets/PathFinding/PathFinding.cs
@@ -1,24 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.AI;
-
-public class PathFinding : MonoBehaviour
-{
-    public Transform goal;
-    private NavMeshAgent agent;
-    void Start()
-    {
-        agent = GetComponent<NavMeshAgent>();
-        agent.updateRotation = false;
-        agent.updateUpAxis = false;
-    }
-
-    private void Update()
-    {
-        if (Input.GetKeyDown(KeyCode.Space)) {
-            Debug.Log("PRESSED");
-            agent.destination = goal.position;
-        }
-    }
-}
